Refactor code to use React Hooks, add Routing and clean component hierarchy in the project
Refactored the code to use React Hooks. Also refactored the photobrowser to only display the album and thumbnail infos and the clicked modal has it's own container with photodisplayer which handles displaying the clicked image and showing it's info. Also added routing to the gallery images so showing a large image is possible with certain url.